bot-core/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=1.10.13
pydantic-settings>=2.0.0
aiokafka>=0.10.0
aiogram>=3.4.1
httpx>=0.27.0
kafka-python>=2.0.2
requests>=2.31.0
beautifulsoup4>=4.12.0
python-docx>=1.1.0
pdfminer.six>=20221105

